Updated macOS script to Linux state
Updated macOS script to contain all features of the Linux version. Also corrected an error in the README.md file.

+++ osx/g2dump.sh
@@ -1,13 +1,18 @@
 #!/bin/bash
 
+START=${1:-"2014-10-27"}
+TODAY=$2
+
 OPTIND=1
 
 FROM="2014-10-27"
 TO=$(date -j -f "%Y%m%d" "$(date "+%Y%m%d")" "+%s")
 OUTPUT="./"
 SOURCE="http://www.galcon.com/g2/logs/"
+COMPRESS=false
+DOWNLOAD_COMPRESSED=false
 
-while getopts "f:t:o:s:" OPT; do
+while getopts "f:t:o:s:cx" OPT; do
   case "$OPT" in
     f) #from
       FROM=$OPTARG
@@ -21,6 +26,12 @@ while getopts "f:t:o:s:" OPT; do
     s) #source
       SOURCE=$OPTARG
       ;;
+    c) #enable compression
+      COMPRESS=true
+      ;;
+    x) #try to download compressed .tgz versions of the dump (and extract them)
+      DOWNLOAD_COMPRESSED=true
+      ;;
   esac
 done
 
@@ -28,7 +39,28 @@ while [ $(date -j -f "%Y-%m-%d" "$FROM" "+%s") -le $TO ]; do
   if [ -f $OUTPUT$FROM.txt ]; then
     echo $FROM.txt already exists. Skipping.
   else
-    curl "$SOURCE$FROM.txt" -f -o "$OUTPUT$FROM.txt"
+    if [ $DOWNLOAD_COMPRESSED = true ]; then
+      if [ ! -f $OUTPUT$FROM.txt ]; then
+        curl "$SOURCE$FROM.tgz" -f -o "$OUTPUT$FROM.tgz"
+      fi
+
+      if [ -f $OUTPUT$FROM.tgz ]; then
+        tar -C "$OUTPUT" -zxvf "$OUTPUT$FROM.tgz"
+      else
+        curl "$SOURCE$FROM.txt" -f -o "$OUTPUT$FROM.txt"
+      fi
+    else
+      curl "$SOURCE$FROM.txt" -f -o "$OUTPUT$FROM.txt"
+    fi
   fi
+
+  if [ $COMPRESS = true ]; then
+    if [ -f $OUTPUT$FROM.tgz ]; then
+      echo $FROM.tgz already exists. Skipping.
+    elif [ -f $OUTPUT$FROM.txt ]; then
+      tar -C "$OUTPUT" -zcvf "$OUTPUT$FROM.tgz" "$FROM.txt"
+    fi
+  fi
+
   FROM=$(date -v+1d -j -f "%Y-%m-%d" "$FROM" "+%Y-%m-%d")
 done
